Show the variation picker again when it's inner-blocks are removed
